requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.0
torch>=2.0.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
googletrans==4.0.0-rc1
gtts>=2.4.0
numpy>=1.24.0
huggingface-hub>=0.19.0
accelerate>=0.24.0

